+++ regression_tests/actions/create_virtual_environment.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+set -e
+
+echo -e "\n\n### Create virtual environment\n"
+
+source ./regression_tests/src/venv_helper.sh
+
+create_and_activate_venv