+++ regression_tests/actions/ensure_not_under_maintenance.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-echo -e "\n\nEnsure environment not under maintenance\n"
+echo -e "\n\n### Ensure ${TARGET_ENVIRONMENT} environment not under maintenance\n"
 
 cd "${CODEBUILD_SRC_DIR}/demodjango"