+++ regression_tests/actions/run_maintenance_page_tests.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-echo -e "\n\nRunning maintenance page tests\n"
+echo -e "\n\n### Run maintenance page tests for ${TARGET_ENVIRONMENT} environment \n"
 
 cd "${CODEBUILD_SRC_DIR}/demodjango-deploy"
 echo "Current demodjango-deploy branch/commit: $(git rev-parse --abbrev-ref HEAD)/$(git rev-parse HEAD)"
@@ -11,6 +11,7 @@ URL=https://internal.${TARGET_ENVIRONMENT}.demodjango.uktrade.digital/
 
 echo -e "\nCheck we can view the page (running smoke tests)"
 cd "${CODEBUILD_SRC_DIR}/demodjango"
+# Todo: We should probably tighten this up to just run the tests we are interest in...
 ./smoke_tests.sh ${TARGET_ENVIRONMENT} smoke
 
 echo -e "\nRunning offline command"