+++ regression_tests/actions/set_up_aws_config.sh
@@ -2,12 +2,13 @@
 
 set -e
 
-echo -e "\n\nAssume platform-sandbox role to trigger environment pipeline\n"
+echo -e "\n\n### Assume platform-sandbox role to trigger environment pipeline\n"
 
 cd "${CODEBUILD_SRC_DIR}"
 
+# Todo: Look to extract to a helper script in a followup pull request
 assumed_role=$(aws sts assume-role \
-    --role-arn "arn:aws:iam::$PLATFORM_SANDBOX_AWS_ACCOUNT_ID:role/regression-tests-assume-role-for-platform-tools" \
+    --role-arn "arn:aws:iam::$PLATFORM_SANDBOX_AWS_ACCOUNT_ID:role/platform-regression-tests-role" \
     --role-session-name "pull-request-regression-tests-$(date +%s)")
 PLATFORM_SANDBOX_AWS_ACCESS_KEY_ID=$(echo $assumed_role | jq -r .Credentials.AccessKeyId)
 PLATFORM_SANDBOX_AWS_SECRET_ACCESS_KEY=$(echo $assumed_role | jq -r .Credentials.SecretAccessKey)