+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [10.11.3](https://github.com/uktrade/platform-tools/compare/10.11.2...10.11.3) (2024-10-02)
+
+
+### Bug Fixes
+
+* DBTP-1373 - Handle SchemaErrors from platform config validation gracefully ([#584](https://github.com/uktrade/platform-tools/issues/584)) ([8324049](https://github.com/uktrade/platform-tools/commit/83240491fede3130c5c4d338029bb139bf7d2ef0))
+* DBTP-1375 - platform helper version get-platform-helper-for-project no longer validates config ([#575](https://github.com/uktrade/platform-tools/issues/575)) ([f62c25d](https://github.com/uktrade/platform-tools/commit/f62c25d5fcb218c1668be1a7043ca45745c7c03b))
+* fix typo ([#591](https://github.com/uktrade/platform-tools/issues/591)) ([8f48dd3](https://github.com/uktrade/platform-tools/commit/8f48dd3dc6449045add87215483a625290b21484))
+
+## [10.11.2](https://github.com/uktrade/platform-tools/compare/10.11.1...10.11.2) (2024-09-26)
+
+
+### Bug Fixes
+
+* DBTP-1382 Fix conduit session KMS access permissions ([#583](https://github.com/uktrade/platform-tools/issues/583)) ([fc3051e](https://github.com/uktrade/platform-tools/commit/fc3051e4c9591a0a87cc28dab7f413ca4acde9c1))
+* Restrict regression test alerts to toolspr environment ([#581](https://github.com/uktrade/platform-tools/issues/581)) ([d40f50f](https://github.com/uktrade/platform-tools/commit/d40f50fb9c8b4c552f223e97b31454e3afc302d7))
+
 ## [10.11.1](https://github.com/uktrade/platform-tools/compare/10.11.0...10.11.1) (2024-09-24)